trying to debug

+++ providers/apps.rb
@@ -50,10 +50,13 @@ action :remove do
   query = "NOT (#{node[:roles].map{|r| "roles:#{r}" }.join(" OR ")})"
   Chef::Log.info("Running query: #{query}")
   search("#{new_resource.data_bag}", "#{query}") do |site|
-    deploy_config = site[:deploy]
-    deploy_config[:home] ||= new_resource.home
-    deploy_config[:deploy_to] ||= File.join(deploy_config[:home], site[:id])
-    execute("rm -rf #{deploy_config[:deploy_to]}")
+    Chef::Log.info("----------------------------------------------")
+    Chef::Log.info(site)
+    if deploy_config = site[:deploy]
+      deploy_config[:home] ||= new_resource.home
+      deploy_config[:deploy_to] ||= File.join(deploy_config[:home], site[:id])
+      execute("rm -rf #{deploy_config[:deploy_to]}")
+    end
   end
 end
